修复处理文件时 `Uint32Array` 报错的问题

+++ src/swpp/database/CompilationFileParser.ts
@@ -44,7 +44,7 @@ export class CompilationFileParser extends KeyValueDatabase<FileParser<crypto.Bi
         } else {
             if (callback) {
                 const buffer = await response.arrayBuffer()
-                const array = new Uint32Array(buffer)
+                const array = new Uint8Array(buffer)
                 callback(array)
             }
             return new Set<string>()
